requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.2
flask_cors==4.0.0
sqlalchemy==2.0.21
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.4
pyjwt==2.8.0
